move permissions section to a tab
